11 issues updated

+++ WebVella.Erp.Plugins.Mail/MailPlugin.20190419.cs
@@ -278,7 +278,7 @@ namespace WebVella.Erp.Plugins.Mail
 				var url = "/sdk/access/user/l/list";
 				var iconClass = "fa fa-user";
 				var weight = 1;
-				var type = ((int)3);
+				var type = 3;
 				var access = new List<Guid>();
 				var entityListPages = new List<Guid>();
 				var entityCreatePages = new List<Guid>();
